+++ src/smart_follower_control/include/smart_follower_control/constants.hpp
@@ -1,9 +1,9 @@
-﻿#pragma once
+#pragma once
 
 namespace smart_follower_control
 {
 
-inline constexpr char kRuntimeVersion[] = "beta-0.3.0";
+inline constexpr char kRuntimeVersion[] = "beta-0.3.1";
 
 }  // namespace smart_follower_control